10/03/23 Added the Season Control.
The Season Control is intended to work as a reference for seasons and it's vital for the Weather Machine.

It's a numeric slider that goes from -42 to 42. - It has 2 methods:

* Tick(): Moves 1 day forward and turns around if needed.
* CustomTick(): moves forward (positive increment) or backward (negative increment) and turns around if needed.

By "turning around" I mean changing the movingForward variable, which indicates if the seasons are either moving towards the peak of Summer or the peak of Winter.