Update Weather.php
- Created a new class: Season Control

This class is a temporary receptacle for the new table SeasonControl, or either the fields Day and GoingForward on an existing table.
These fields hold the value that represent together the Seasons cycle, giving the direction (if the cycle is moving towards the warm or the cold season) and the day at the cycle.
It also has a method that returns the current season based on those values. There are also 2 other methods: Tick() and CustomTick().

* Tick(): moves 1 day forward on the Seasons cycle. When it reaches either end, it turns around changing the GoingForward variable.

* CustomTick(): moves N days on the Seasons cycle. If the number is positive, the movement is in the forward direction. If it's negative, the movement is backwards.

- Created a new class: WeatherSystemDataAccess

This class receives de DB Path as a string and is in charge of reading the DB tables and returning objects, and updating entries on the DB corresponding to objects.
The methods are currently and temporarily written for their use with a SQLite DB. I will soon modify them in order to make them work with a SQL/MySQL database when I make it home and I get some time.